JSON methods cleanup.
